refactor actor/action views, moving compleexity to sub-components
