+++ app/containers/ActionView/index.js
@@ -10,8 +10,7 @@ import { connect } from 'react-redux';
 import Helmet from 'react-helmet';
 import { List } from 'immutable';
 import { FormattedMessage, injectIntl, intlShape } from 'react-intl';
-import { Box, Text, Button } from 'grommet';
-import styled from 'styled-components';
+import { Box } from 'grommet';
 
 import {
   getTitleField,
@@ -26,7 +25,6 @@ import {
   getNumberField,
   getTaxonomyFields,
   hasTaxonomyCategories,
-  getActorConnectionField,
   getActionConnectionField,
   getResourceConnectionField,
 } from 'utils/fields';
@@ -41,17 +39,19 @@ import {
   loadEntitiesIfNeeded,
   updatePath,
   closeEntity,
-  setSubject,
 } from 'containers/App/actions';
 
 import {
+  selectReady,
+  selectIsUserManager,
+  selectResourceConnections,
+  selectTaxonomiesWithCategories,
+} from 'containers/App/selectors';
+
+import {
   ROUTES,
-  ACTIONTYPES,
   FF_ACTIONTYPE,
-  ACTORTYPES_CONFIG,
-  ACTORTYPES,
   RESOURCE_FIELDS,
-  ACTIONTYPE_ACTOR_ACTION_ROLES,
 } from 'themes/config';
 
 import Loading from 'components/Loading';
@@ -63,123 +63,22 @@ import ViewWrapper from 'components/EntityView/ViewWrapper';
 import ViewPanel from 'components/EntityView/ViewPanel';
 import ViewPanelInside from 'components/EntityView/ViewPanelInside';
 import FieldGroup from 'components/fields/FieldGroup';
-import ButtonDefault from 'components/buttons/ButtonDefault';
 
-import {
-  selectReady,
-  selectIsUserManager,
-  selectActorConnections,
-  selectResourceConnections,
-  selectTaxonomiesWithCategories,
-  selectSubjectQuery,
-  selectActiontypes,
-} from 'containers/App/selectors';
 
 import appMessages from 'containers/App/messages';
 import messages from './messages';
 
-import ActionMap from './ActionMap';
-import IndicatorCountryMap from './IndicatorCountryMap';
-import IndicatorLocationMap from './IndicatorLocationMap';
+import Actors from './Actors';
+
 import {
   selectViewEntity,
   selectViewTaxonomies,
-  selectActorsByType,
-  selectTargetsByType,
   selectResourcesByType,
   selectChildActionsbyType,
   selectParentAction,
 } from './selectors';
 
 import { DEPENDENCIES } from './constants';
-
-const SubjectButton = styled((p) => <Button plain {...p} />)`
-  padding: 2px 4px;
-  border-bottom: 2px solid;
-  border-bottom-color: ${({ active }) => active ? 'brand' : 'transparent'};
-  background: none;
-`;
-
-const getActortypeColumns = (
-  actortypeId,
-  isIndicator,
-  isActive,
-  viewEntity,
-  intl,
-) => {
-  let columns = [{
-    id: 'main',
-    type: 'main',
-    sort: 'title',
-    attributes: ['code', 'title'],
-    isIndicator,
-  }];
-  if (qe(actortypeId, ACTORTYPES.COUNTRY)) {
-    columns = [
-      ...columns,
-      {
-        id: 'classes',
-        type: 'associations',
-        actortype_id: ACTORTYPES.CLASS,
-        title: 'Classes',
-        isIndicator,
-      },
-    ];
-    if (isIndicator) {
-      columns = [
-        ...columns,
-        {
-          id: 'regions',
-          type: 'associations',
-          actortype_id: ACTORTYPES.REG,
-          title: 'Regions',
-          isIndicator,
-        },
-      ];
-    }
-  }
-  if (isIndicator) {
-    columns = [
-      ...columns,
-      {
-        id: 'indicator',
-        type: 'indicator',
-        indicatorId: viewEntity.get('id'),
-        title: viewEntity.getIn(['attributes', 'title']),
-        unit: viewEntity.getIn(['attributes', 'comment']),
-        align: 'end',
-        primary: true,
-      },
-    ];
-  }
-  if (
-    ACTORTYPES_CONFIG[parseInt(actortypeId, 10)]
-    && ACTORTYPES_CONFIG[parseInt(actortypeId, 10)].columns
-  ) {
-    columns = [
-      ...columns,
-      ...ACTORTYPES_CONFIG[parseInt(actortypeId, 10)].columns,
-    ];
-  }
-
-  // relationship type
-  if (
-    isActive
-    && ACTIONTYPE_ACTOR_ACTION_ROLES[viewEntity.getIn(['attributes', 'measuretype_id'])]
-    && ACTIONTYPE_ACTOR_ACTION_ROLES[viewEntity.getIn(['attributes', 'measuretype_id'])].length > 0
-  ) {
-    columns = [
-      ...columns,
-      {
-        id: 'relationshiptype_id',
-        type: 'relationship',
-        actionId: viewEntity.get('id'),
-        title: intl.formatMessage(appMessages.attributes.relationshiptype_id),
-      },
-    ];
-  }
-  return columns;
-};
 
 export function ActionView(props) {
   const {
@@ -188,23 +87,16 @@ export function ActionView(props) {
     isManager,
     taxonomies,
     viewTaxonomies,
-    actorsByActortype,
-    targetsByActortype,
     resourcesByResourcetype,
     onEntityClick,
-    actorConnections,
     resourceConnections,
     childrenByType,
     parent,
     onLoadData,
-    subject,
-    onSetSubject,
     intl,
     handleEdit,
     handleClose,
     params,
-    activitytypes,
-    handleImportConnection,
     handleTypeClick,
   } = props;
 
@@ -214,8 +106,8 @@ export function ActionView(props) {
   }, []);
 
   const typeId = viewEntity && viewEntity.getIn(['attributes', 'measuretype_id']);
-  const viewActivitytype = activitytypes && activitytypes.find((type) => qe(type.get('id'), typeId));
   const isIndicator = qe(typeId, FF_ACTIONTYPE);
+
   let buttons = [];
   if (dataReady) {
     buttons = [
@@ -245,18 +137,6 @@ export function ActionView(props) {
     ? `${pageTitle}: ${getEntityTitleTruncated(viewEntity)}`
     : `${pageTitle}: ${params.id}`;
 
-  const hasTarget = viewActivitytype && viewActivitytype.getIn(['attributes', 'has_target']);
-  const hasMemberOption = !!typeId && !qe(typeId, ACTIONTYPES.NATL);
-  const viewSubject = hasTarget && subject ? subject : 'actors';
-
-  const actortypesForSubject = !hasTarget || viewSubject === 'actors'
-    ? actorsByActortype
-    : targetsByActortype;
-
-  // action has a map
-  const hasCountryActionMap = !!typeId && !isIndicator;
-  const hasIndicatorCountryMap = !!typeId && isIndicator && actortypesForSubject && actortypesForSubject.get(parseInt(ACTORTYPES.COUNTRY, 10));
-  const hasIndicatorLocationMap = !!typeId && isIndicator && actortypesForSubject && actortypesForSubject.get(parseInt(ACTORTYPES.POINT, 10));
   // && !qe(typeId, ACTIONTYPES.NATL);
   let hasLandbasedValue;
   if (viewEntity && checkActionAttribute(typeId, 'has_reference_landbased_ml')) {
@@ -386,162 +266,58 @@ export function ActionView(props) {
                       ],
                     }}
                   />
+                  <Actors
+                    id={params.id}
+                    viewEntity={viewEntity}
+                    dataReady={dataReady}
+                    isManager={isManager}
+                    taxonomies={taxonomies}
+                    onEntityClick={onEntityClick}
+                    typeId={typeId}
+                    isIndicator={isIndicator}
+                  />
                   <Box>
-                    {!isIndicator && (
-                      <Box direction="row" gap="small" margin={{ vertical: 'small', horizontal: 'medium' }}>
-                        <SubjectButton
-                          onClick={() => onSetSubject('actors')}
-                          active={viewSubject === 'actors'}
-                        >
-                          <Text size="large">{qe(ACTIONTYPES.DONOR, typeId) ? 'Donors' : 'Actors'}</Text>
-                        </SubjectButton>
-                        {hasTarget && (
-                          <SubjectButton
-                            onClick={() => onSetSubject('targets')}
-                            active={viewSubject === 'targets'}
-                          >
-                            <Text size="large">{qe(ACTIONTYPES.DONOR, typeId) ? 'Recipients' : 'Targets'}</Text>
-                          </SubjectButton>
-                        )}
-                      </Box>
-                    )}
-                    {(!actortypesForSubject || actortypesForSubject.size === 0) && (
-                      <Box margin={{ vertical: 'small', horizontal: 'medium' }}>
-                        {viewSubject === 'actors' && (
-                          <Text>
-                            No actors for activity in database
-                          </Text>
-                        )}
-                        {viewSubject === 'targets' && (
-                          <Text>
-                            No activity targets in database
-                          </Text>
-                        )}
-                      </Box>
-                    )}
-                    <Box>
-                      {dataReady && actortypesForSubject && hasCountryActionMap && (
-                        <ActionMap
-                          entities={actortypesForSubject}
-                          mapSubject={viewSubject}
-                          onActorClick={(id) => onEntityClick(id, ROUTES.ACTOR)}
-                          hasMemberOption={hasMemberOption}
-                          typeId={typeId}
-                        />
-                      )}
-                      {dataReady && actortypesForSubject && hasIndicatorCountryMap && (
-                        <IndicatorCountryMap
-                          countries={actortypesForSubject.get(parseInt(ACTORTYPES.COUNTRY, 10))}
-                          mapSubject="actors"
-                          onCountryClick={(id) => onEntityClick(id, ROUTES.ACTOR)}
-                          indicator={viewEntity}
-                        />
-                      )}
-                      {dataReady && actortypesForSubject && hasIndicatorLocationMap && (
-                        <IndicatorLocationMap
-                          locations={actortypesForSubject.get(parseInt(ACTORTYPES.POINT, 10))}
-                          mapSubject="actors"
-                          indicator={viewEntity}
-                        />
-                      )}
-                      {viewSubject === 'targets' && hasTarget && (
-                        <FieldGroup
-                          group={{
-                            fields: [
-                              checkActionAttribute(typeId, 'target_comment')
-                                && getMarkdownField(viewEntity, 'target_comment', true),
-                            ],
-                          }}
-                        />
-                      )}
-                      {actortypesForSubject && (
-                        <FieldGroup
-                          group={{
-                            fields: actortypesForSubject.reduce(
-                              (memo, actors, typeid) => memo.concat([
-                                getActorConnectionField({
-                                  actors: isIndicator
-                                    ? actors.filter(
-                                      (actor) => !!actor.getIn(['actionValues', viewEntity.get('id')])
-                                    )
-                                    : actors,
-                                  taxonomies,
-                                  onEntityClick,
-                                  connections: actorConnections,
-                                  typeid,
-                                  columns: getActortypeColumns(
-                                    typeid,
-                                    isIndicator,
-                                    viewSubject === 'actors',
-                                    viewEntity,
-                                    intl,
-                                  ),
-                                  isIndicator,
-                                  sortBy: isIndicator ? 'indicator' : null,
-                                  sortOrder: isIndicator ? 'desc' : null,
-                                }),
-                              ]),
-                              [],
-                            ),
-                          }}
-                        />
-                      )}
-                      {isManager && isIndicator && (
-                        <Box
-                          margin={{ bottom: 'large', horizontal: 'medium' }}
-                          fill={false}
-                          alignContent="start"
-                          direction="row"
-                        >
-                          <ButtonDefault
-                            onClick={() => handleImportConnection()}
-                          >
-                            Import actor connections
-                          </ButtonDefault>
-                        </Box>
-                      )}
-                      {resourcesByResourcetype && (
-                        <FieldGroup
-                          group={{
-                            type: 'dark',
-                            label: appMessages.nav.resources,
-                            fields: resourcesByResourcetype.reduce(
-                              (memo, resources, resourcetypeid) => {
-                                let columns = [
+                    {resourcesByResourcetype && (
+                      <FieldGroup
+                        group={{
+                          type: 'dark',
+                          label: appMessages.nav.resources,
+                          fields: resourcesByResourcetype.reduce(
+                            (memo, resources, resourcetypeid) => {
+                              let columns = [
+                                {
+                                  id: 'main',
+                                  type: 'main',
+                                  sort: 'title',
+                                  attributes: ['title'],
+                                },
+                              ];
+                              if (RESOURCE_FIELDS.ATTRIBUTES.status.optional.indexOf(resourcetypeid.toString()) > -1) {
+                                columns = [
+                                  ...columns,
                                   {
-                                    id: 'main',
-                                    type: 'main',
-                                    sort: 'title',
-                                    attributes: ['title'],
+                                    id: 'attribute',
+                                    type: 'attribute',
+                                    attribute: 'status',
                                   },
                                 ];
-                                if (RESOURCE_FIELDS.ATTRIBUTES.status.optional.indexOf(resourcetypeid.toString()) > -1) {
-                                  columns = [
-                                    ...columns,
-                                    {
-                                      id: 'attribute',
-                                      type: 'attribute',
-                                      attribute: 'status',
-                                    },
-                                  ];
-                                }
-                                return memo.concat([
-                                  getResourceConnectionField({
-                                    resources,
-                                    taxonomies,
-                                    onEntityClick,
-                                    connections: resourceConnections,
-                                    typeid: resourcetypeid,
-                                    columns,
-                                  }),
-                                ]);
-                              },
-                              [],
-                            ),
-                          }}
-                        />
-                      )}
-                    </Box>
+                              }
+                              return memo.concat([
+                                getResourceConnectionField({
+                                  resources,
+                                  taxonomies,
+                                  onEntityClick,
+                                  connections: resourceConnections,
+                                  typeid: resourcetypeid,
+                                  columns,
+                                }),
+                              ]);
+                            },
+                            [],
+                          ),
+                        }}
+                      />
+                    )}
                   </Box>
                 </Main>
                 {!isIndicator && (
@@ -653,25 +429,13 @@ ActionView.propTypes = {
   isManager: PropTypes.bool,
   viewTaxonomies: PropTypes.object,
   taxonomies: PropTypes.object,
-  actorsByActortype: PropTypes.object,
-  targetsByActortype: PropTypes.object,
   resourcesByResourcetype: PropTypes.object,
-  actorConnections: PropTypes.object,
   resourceConnections: PropTypes.object,
-  activitytypes: PropTypes.object,
   params: PropTypes.object,
   childrenByType: PropTypes.object,
   parent: PropTypes.object,
-  onSetSubject: PropTypes.func,
-  handleImportConnection: PropTypes.func,
   intl: intlShape.isRequired,
-  subject: PropTypes.string,
 };
-
-ActionView.contextTypes = {
-  intl: PropTypes.object.isRequired,
-};
-
 
 const mapStateToProps = (state, props) => ({
   isManager: selectIsUserManager(state),
@@ -679,15 +443,10 @@ const mapStateToProps = (state, props) => ({
   viewEntity: selectViewEntity(state, props.params.id),
   viewTaxonomies: selectViewTaxonomies(state, props.params.id),
   taxonomies: selectTaxonomiesWithCategories(state),
-  actorsByActortype: selectActorsByType(state, props.params.id),
   resourcesByResourcetype: selectResourcesByType(state, props.params.id),
-  targetsByActortype: selectTargetsByType(state, props.params.id),
-  actorConnections: selectActorConnections(state),
   resourceConnections: selectResourceConnections(state),
   childrenByType: selectChildActionsbyType(state, props.params.id),
   parent: selectParentAction(state, props.params.id),
-  subject: selectSubjectQuery(state),
-  activitytypes: selectActiontypes(state),
 });
 
 function mapDispatchToProps(dispatch, props) {
@@ -701,17 +460,11 @@ function mapDispatchToProps(dispatch, props) {
     handleEdit: () => {
       dispatch(updatePath(`${ROUTES.ACTION}${ROUTES.EDIT}/${props.params.id}`, { replace: true }));
     },
-    handleImportConnection: () => {
-      dispatch(updatePath(`${ROUTES.ACTOR_ACTIONS}${ROUTES.IMPORT}/${props.params.id}`, { replace: true }));
-    },
     handleClose: (typeId) => {
       dispatch(closeEntity(`${ROUTES.ACTIONS}/${typeId}`));
     },
     handleTypeClick: (typeId) => {
       dispatch(updatePath(`${ROUTES.ACTIONS}/${typeId}`));
-    },
-    onSetSubject: (type) => {
-      dispatch(setSubject(type));
     },
   };
 }
